refactor: `SmsOutbox` 관련 클래스 도메인에서 예약 SMS 전용 패키지로 이동
- `SmsOutbox`, `SmsOutboxEventType`, `SmsOutboxStatus` 클래스를 `fittoring.application.reservation.sms`로 이동
- 관련된 모든 참조 경로를 새로운 패키지 경로로 수정
- 주요 테스트 코드 및 서비스 로직에서 변경된 경로 반영
