+++ backend/fittoring/src/main/java/fittoring/admin/service/AdminSmsOutboxService.java
@@ -5,8 +5,8 @@ import fittoring.admin.presentation.dto.AdminSmsOutboxResponse;
 import fittoring.admin.presentation.dto.PageResult;
 import fittoring.application.exception.SmsOutboxNotFoundException;
 import fittoring.application.reservation.repository.SmsOutboxRepository;
-import fittoring.domain.model.SmsOutbox;
-import fittoring.domain.model.SmsOutboxStatus;
+import fittoring.application.reservation.sms.SmsOutbox;
+import fittoring.application.reservation.sms.SmsOutboxStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
