+++ backend/fittoring/src/test/java/fittoring/integration/SmsOutboxClaimIntegrationTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import fittoring.AbstractApiDocumentationTest;
 import fittoring.application.reservation.repository.SmsOutboxRepository;
 import fittoring.domain.model.Phone;
-import fittoring.domain.model.SmsOutbox;
-import fittoring.domain.model.SmsOutboxEventType;
-import fittoring.domain.model.SmsOutboxStatus;
+import fittoring.application.reservation.sms.SmsOutbox;
+import fittoring.application.reservation.sms.SmsOutboxEventType;
+import fittoring.application.reservation.sms.SmsOutboxStatus;
 import fittoring.infrastructure.SmsOutboxClaimer;
 import java.time.LocalDateTime;
 import java.util.List;
