+++ backend/fittoring/src/test/java/fittoring/integration/SmsOutboxDiscordNotificationIntegrationTest.java
@@ -10,9 +10,9 @@ import static org.mockito.Mockito.verifyNoInteractions;
 import fittoring.AbstractApiDocumentationTest;
 import fittoring.application.reservation.repository.SmsOutboxRepository;
 import fittoring.domain.model.Phone;
-import fittoring.domain.model.SmsOutbox;
-import fittoring.domain.model.SmsOutboxEventType;
-import fittoring.domain.model.SmsOutboxStatus;
+import fittoring.application.reservation.sms.SmsOutbox;
+import fittoring.application.reservation.sms.SmsOutboxEventType;
+import fittoring.application.reservation.sms.SmsOutboxStatus;
 import fittoring.infrastructure.SmsOutboxPublisher;
 import fittoring.infrastructure.SmsOutboxResultApplier;
 import fittoring.infrastructure.exception.InfraErrorMessage;
