+++ backend/fittoring/src/test/java/fittoring/integration/SmsOutboxPublisherIntegrationTest.java
@@ -8,9 +8,9 @@ import static org.mockito.Mockito.doThrow;
 import fittoring.AbstractApiDocumentationTest;
 import fittoring.application.reservation.repository.SmsOutboxRepository;
 import fittoring.domain.model.Phone;
-import fittoring.domain.model.SmsOutbox;
-import fittoring.domain.model.SmsOutboxEventType;
-import fittoring.domain.model.SmsOutboxStatus;
+import fittoring.application.reservation.sms.SmsOutbox;
+import fittoring.application.reservation.sms.SmsOutboxEventType;
+import fittoring.application.reservation.sms.SmsOutboxStatus;
 import fittoring.infrastructure.SmsOutboxPublisher;
 import fittoring.infrastructure.SmsOutboxResultApplier;
 import fittoring.infrastructure.dto.BatchSendResult;
