+++ backend/fittoring/src/test/java/fittoring/integration/admin/AdminSmsOutboxIntegrationTest.java
@@ -13,9 +13,9 @@ import fittoring.domain.model.Gender;
 import fittoring.domain.model.Member;
 import fittoring.domain.model.MemberRole;
 import fittoring.domain.model.Phone;
-import fittoring.domain.model.SmsOutbox;
-import fittoring.domain.model.SmsOutboxEventType;
-import fittoring.domain.model.SmsOutboxStatus;
+import fittoring.application.reservation.sms.SmsOutbox;
+import fittoring.application.reservation.sms.SmsOutboxEventType;
+import fittoring.application.reservation.sms.SmsOutboxStatus;
 import fittoring.domain.model.password.Password;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
